Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,7 +7,7 @@ char* welcome() {
 int main() {
     float a ,b ,c;
     scanf("%f %f %f" , &a,&b,&c);
-    int d = (a+b+c)/3;
+    float d = (a+b+c)/3;
     printf("Average: %.2f", d);
     return 0;
 }
